third_party/WebKit/LayoutTests/fast/css-generated-content/pseudo-transition-event-expected.txt

Transitions on :before and :after pseudo elements should run and fire DOM events

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS event.propertyName is expectedEvents[currentEvent][0]
PASS event.target.id is expectedEvents[currentEvent][1]
PASS event.pseudoElement is expectedEvents[currentEvent][2]
PASS event.propertyName is expectedEvents[currentEvent][0]
PASS event.target.id is expectedEvents[currentEvent][1]
PASS event.pseudoElement is expectedEvents[currentEvent][2]
PASS successfullyParsed is true

TEST COMPLETE

 
